requirements.txt
langchain
langchain_community
langchain_openai
langchain_xai
pydantic-ai
redis
redis-async
uvicorn
fastapi
SQLAlchemy
alembic
python-dotenv
pydantic
pydantic[email]
psycopg2
passlib
PyJWT
celery
python-multipart
google-cloud-storage
pytest
httpx
python-dotenv
websockets
watchdog>=2.0.0
huggingface_hub
diffusers
google-cloud-storage
replicate
sendgrid
pillow
stripe
jinja2